cart, point, stock 대폭 수정

+++ DA-GWAN/src/main/java/hyper/darye/controller/CartController.java
@@ -1,10 +1,13 @@
 package hyper.darye.controller;
 
-import hyper.darye.dto.CartSelect;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import hyper.darye.dto.controller.request.SelectCartRequest;
 import hyper.darye.service.CartService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.ArrayList;
 import java.util.List;
 
 @RestController
@@ -14,20 +17,29 @@ public class CartController {
     @Autowired
     private CartService cartService;
 
+    @Autowired
+    private ObjectMapper objectMapper;
+
     // 장바구니 넣기
-    @PostMapping("/{id}/cart")
-    public String addCart(@PathVariable Long id, Long productId, Long quantity) {
-        int result = cartService.insertCart(id, productId, quantity);
+    @PreAuthorize("#memberId == authentication.principal.id")
+    @PostMapping("/{memberId}/cart")
+    public String insertCart(@PathVariable Long memberId,
+                             @RequestParam Long productId,
+                             @RequestParam Long quantity) {
+        int result = cartService.insertCart(memberId, productId, quantity);
         if (result == 1)
             return "성공";
         else
-            throw new IllegalStateException("입력 실패"); // 적절한 런타임 예외
+            throw new IllegalStateException("입력 실패했습니다."); // 적절한 런타임 예외
     }
 
     // 장바구니 수량 변경
-    @PatchMapping("/{id}/cart/{productId}")
-    public String updateCart(@PathVariable Long id, Long productId, @RequestParam Long quantity) {
-        int result = cartService.updateCartQuantity(id, productId, quantity);
+    @PreAuthorize("#memberId == authentication.principal.id")
+    @PatchMapping("/{memberId}/cart/{productId}")
+    public String updateCart(@PathVariable Long memberId,
+                             @PathVariable  Long productId,
+                             @RequestParam Long quantity) {
+        int result = cartService.updateCartQuantity(memberId, productId, quantity);
         if (result == 1)
             return "성공";
         else
@@ -36,15 +48,24 @@ public class CartController {
 
 
     // 장바구니 조회
-    @GetMapping("/{id}/cart")
-    public List<CartSelect> searchCart(@PathVariable Long id) {
-        return cartService.selectCart(id);
+    @PreAuthorize("#memberId == authentication.principal.id")
+    @GetMapping("/{memberId}/cart")
+    public List<SelectCartRequest> selectCartByMemberId(@PathVariable Long memberId)
+    {
+        List<SelectCartRequest> searchCart = cartService.selectCart(memberId);
+        if(searchCart.isEmpty()) return new ArrayList<>();
+
+        return cartService.selectCart(memberId);
     }
 
 
     // 장바구니 선택 삭제
-    @DeleteMapping("/{id}/cart")
-    public int deleteCart(Long memberId, @RequestParam List<Long> productIdList) {
+    @PreAuthorize("#memberId == authentication.principal.id")
+    @DeleteMapping("/{memberId}/cart")
+    public int deleteCart(@PathVariable Long memberId, @RequestParam List<Long> productIdList) {
+        if (productIdList == null || productIdList.isEmpty()) {
+            return 0; // 삭제할 항목이 없음을 나타냄
+        }
         return cartService.deleteCart(memberId, productIdList);
     }
 }
